requirements.txt
unsloth
trl
peft
accelerate
qwen_vl_utils
transformers
huggingface
datasets
opencv-python
Pillow
tqdm
json5
ultralytics
seaborn
numpy
pandas
matplotlib
hf_xet
bitsandbytes
einops
safetensors
hf_xet
pyyaml
